requirements.txt
numpy
scipy
h5py
matplotlib
vapory
Pillow
vpython
mayavi
